# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssistantInfo": {
        "properties": {
          "available": {
            "description": "False when no AI provider is configured on the server",
            "title": "Available",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssistantStatus"
          },
          "suggested_questions": {
            "items": {
              "type": "string"
            },
            "title": "Suggested Questions",
            "type": "array"
          }
        },
        "required": [
          "domain",
          "display_name",
          "suggested_questions",
          "available",
          "status"
        ],
        "title": "AssistantInfo",
        "type": "object"
      },
      "AssistantStatus": {
        "description": "What the AI provider's own responses last reported (never guessed). `unconfirmed`: nothing\nknown yet, or a limit's reset time has passed; the next question's real response decides.",
        "properties": {
          "limit_scope": {
            "anyOf": [
              {
                "enum": [
                  "daily",
                  "minute"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Scope"
          },
          "resets_in_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Seconds until questions are accepted again, from the provider's reply",
            "title": "Resets In Seconds"
          },
          "state": {
            "enum": [
              "available",
              "limit_reached",
              "unconfirmed",
              "not_configured"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "AssistantStatus",
        "type": "object"
      },
...
      "ConversationDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "title": "Messages",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "title",
          "created_at",
          "last_message_at",
          "messages"
        ],
        "title": "ConversationDetail",
        "type": "object"
      },
      "ConversationList": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationSummary"
            },
            "title": "Conversations",
            "type": "array"
          }
        },
        "required": [
          "conversations"
        ],
        "title": "ConversationList",
        "type": "object"
      },
      "ConversationSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "title",
          "created_at",
          "last_message_at"
        ],
        "title": "ConversationSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageOut": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data freshness of the figures used",
            "title": "As Of"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "sources": {
            "description": "For \"Based on …\" (assistant)",
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "complete",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "status",
          "created_at"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "Reply": {
        "properties": {
          "assistant_message": {
            "$ref": "#/components/schemas/MessageOut"
          },
          "conversation": {
            "$ref": "#/components/schemas/ConversationSummary"
          },
          "user_message": {
            "$ref": "#/components/schemas/MessageOut"
          }
        },
        "required": [
          "conversation",
          "user_message",
          "assistant_message"
        ],
        "title": "Reply",
        "type": "object"
      },
...
      "SendMessage": {
        "properties": {
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "SendMessage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clario API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/assistant": {
      "get": {
        "operationId": "get_assistant_api_v1_workspaces__workspace_id__connections__connection_id__assistant_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The assistant for this connection",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/assistant/conversations": {
      "get": {
        "operationId": "list_conversations_api_v1_workspaces__workspace_id__connections__connection_id__assistant_conversations_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Your conversations",
        "tags": [
          "assistant"
        ]
      },
      "post": {
        "operationId": "create_conversation_api_v1_workspaces__workspace_id__connections__connection_id__assistant_conversations_post",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a conversation",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/assistant/conversations/{conversation_id}": {
      "delete": {
        "operationId": "delete_conversation_api_v1_workspaces__workspace_id__connections__connection_id__assistant_conversations__conversation_id__delete",
        "parameters": [
          {
            "description": "Conversation id",
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "description": "Conversation id",
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a conversation",
        "tags": [
          "assistant"
        ]
      },
      "get": {
        "operationId": "get_conversation_api_v1_workspaces__workspace_id__connections__connection_id__assistant_conversations__conversation_id__get",
        "parameters": [
          {
            "description": "Conversation id",
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "description": "Conversation id",
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A conversation",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/assistant/conversations/{conversation_id}/messages": {
      "post": {
        "operationId": "send_message_api_v1_workspaces__workspace_id__connections__connection_id__assistant_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "description": "Conversation id",
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "description": "Conversation id",
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Sending too quickly, or the AI usage limit is reached"
          },
          "503": {
            "description": "Assistant unavailable or not configured"
          }
        },
        "summary": "Ask the assistant",
        "tags": [
          "assistant"
        ]
      }
    },
...
  }
}